+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `FilterRuleBuiltInParameterExtension` to support `Filter` with inverted options.
 - Update `ElementIdExtension` to fix `ElementIdValue`.
 - Add `FilterBuiltInCategoryExtension` to support filter by `BuiltInCategory`.
+- Update `InverseRule` in `FilterRuleExtension`.
 ### Tests
 - Add `ricaun.Revit.DB.Tests` to test `Select` extension for `Element` and `ElementType`.
 - Add `RevitFilterExtensionTest` tests for `ElementFilterExtension`.